refactor functions from program to primitives
