+++ primitives/Cargo.toml
@@ -15,3 +15,4 @@ codec = { package = "parity-scale-codec", version = "3", default-features = fals
 hex = "0.4.3"
 alloy-primitives = { version = "0.7.5", features=["serde"] }
 alloy-sol-types = "0.7.5"
+blake2 = "0.10.6"
